Remove `include/grpcpp/impl/codegen/method_handler.h`
