+++ src/compiler/cpp_generator.cc
@@ -144,7 +144,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/client_context.h",
         "grpcpp/impl/codegen/completion_queue.h",
         "grpcpp/impl/codegen/message_allocator.h",
-        "grpcpp/impl/codegen/method_handler.h",
+        "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/proto_utils.h",
         "grpcpp/impl/codegen/rpc_method.h",
         "grpcpp/impl/codegen/server_callback.h",
@@ -1654,7 +1654,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/impl/codegen/client_unary_call.h",
         "grpcpp/support/client_callback.h",
         "grpcpp/impl/codegen/message_allocator.h",
-        "grpcpp/impl/codegen/method_handler.h",
+        "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/rpc_service_method.h",
         "grpcpp/impl/codegen/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
